requirements.txt
Django==2.1.3
djangorestframework==3.9.0
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
